resolved changes
